+++ R/rbgm.R
@@ -13,7 +13,7 @@
 #' 
 #' @section I. Import: 
 #' \tabular{ll}{
-#'  \code{\link{read_bgm}} \tab read directly from a .bgm file  \cr
+#'  \code{\link{bgm}} \tab read directly from a .bgm file  \cr
 #'  \code{\link{read.faces}} \tab read faces in with option to densify vertices \cr
 #' }
 #'