+++ tests/testthat/test-basic-bgm.R
@@ -3,15 +3,15 @@ context("basic-bgm")
 files <- list.files(system.file("extdata/bgm", package = "bgmfiles"), pattern = "bgm$", 
                     full.names = TRUE)
 
-test_that("read_bgm works", {
+test_that("bgmfile works", {
   expect_that(
-  read_bgm(files[1])
+  bgmfile(files[1])
   , is_a("list"))
-  expect_silent(lapply(files, read_bgm))
+  expect_silent(lapply(files, bgmfile))
 })
 
 for (i in seq_along(files)) {
-b <- read_bgm(files[i])
+b <- bgmfile(files[i])
 test_that("conversion to Spatial works", {
           expect_that(boxSpatial(b), is_a("SpatialPolygonsDataFrame"))
           expect_that(faceSpatial(b), is_a("SpatialLinesDataFrame"))