+++ NEWS.md
@@ -1,5 +1,7 @@
 # dev
 
+* new function bgmfile to replace deprecated read_bgm
+
 * new tests and changes to capture Guam_utm1 variant
 
 # rbgm 0.0.2